requirements.txt
moderngl>=5.8.0
numpy>=1.20.0
scipy>=1.7.0
soundfile>=0.10.0
imageio>=2.9.0
pyyaml>=6.0
